// client/src/components/types.ts
export interface ITheme<T>{
    theme:{
        body: T;
        text?: T;
        itemCategory?: T;
        modal?: T;
        btnHistory?: T;
        boxBackground?: T;
        amountHistory?: T;
        toggleBorder?: T;
        background?: T;
        pink?: T;
        green?: T;
        modalText?: T;
        btnShadowHistory?: T;
        asideGround?: T;
    }
}